examples/requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile examples/requirements/dev.in
#
appdirs==1.4.4
    # via virtualenv
asgiref==3.3.4
    # via django
attrs==21.2.0
    # via pytest
babel==2.9.1
    # via -r examples/requirements/common.in
bleach==3.3.0
    # via readme-renderer
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
colorama==0.4.4
    # via twine
coverage[toml]==5.5
    # via
    #   -r examples/requirements/test.in
    #   pytest-cov
cryptography==3.4.7
    # via secretstorage
distlib==0.3.1
    # via virtualenv
django-debug-toolbar==3.2.1
    # via -r examples/requirements/django_3_2.in
django==3.2.3
    # via
    #   -r examples/requirements/django_3_2.in
    #   django-debug-toolbar
docutils==0.17.1
    # via readme-renderer
factory_boy==2.11.1
    # via -r examples/requirements/test.in
faker==8.1.4
    # via
    #   -r examples/requirements/test.in
    #   factory-boy
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
idna==2.10
    # via requests
importlib-metadata==4.0.1
    # via
    #   click
    #   keyring
    #   pep517
    #   pluggy
    #   pytest
    #   tox
    #   twine
    #   virtualenv
importlib-resources==5.1.3
    # via virtualenv
iniconfig==1.1.1
    # via pytest
jeepney==0.6.0
    # via
    #   keyring
    #   secretstorage
keyring==23.0.1
    # via twine
packaging==20.9
    # via
    #   bleach
    #   pytest
    #   tox
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r examples/requirements/deployment.in
pkginfo==1.7.0
    # via twine
pluggy==0.13.1
    # via
    #   pytest
    #   tox
py==1.10.0
    # via
    #   -r examples/requirements/test.in
    #   pytest
    #   tox
pycparser==2.20
    # via cffi
pygments==2.9.0
    # via readme-renderer
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.0
    # via -r examples/requirements/test.in
pytest-django==4.3.0
    # via -r examples/requirements/test.in
pytest-ordering==0.6
    # via -r examples/requirements/test.in
pytest-pythonpath==0.7.3
    # via -r examples/requirements/test.in
pytest==6.2.4
    # via
    #   -r examples/requirements/test.in
    #   pytest-cov
    #   pytest-django
    #   pytest-ordering
    #   pytest-pythonpath
python-dateutil==2.8.1
    # via faker
pytz==2021.1
    # via
    #   babel
    #   django
readme-renderer==29.0
    # via twine
requests-toolbelt==0.9.1
    # via twine
requests==2.25.1
    # via
    #   requests-toolbelt
    #   twine
rfc3986==1.5.0
    # via twine
secretstorage==3.3.1
    # via keyring
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
    #   readme-renderer
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   coverage
    #   pep517
    #   pytest
    #   tox
tox==3.23.1
    # via -r examples/requirements/test.in
tqdm==4.60.0
    # via twine
twine==3.4.1
    # via -r examples/requirements/deployment.in
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   importlib-metadata
urllib3==1.26.4
    # via requests
virtualenv==20.4.6
    # via tox
webencodings==0.5.1
    # via bleach
zipp==3.4.1
    # via
    #   importlib-metadata
    #   importlib-resources
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
